+++ changelog.md
@@ -3,9 +3,15 @@
 # TODO
 
 * Add round robbin to bamboo bee + fix icon
-* Fix bamboo bee animation
+* Fix bamboo bee animation and sounds
 * Finish tunnel digger
-  * Add upgrades
-  * Add sounds
+    * Add sounds
+      * On/Off drill
+      * Fix volume via new IA update
 * Finish copperfin
+    * Swim
+    * Fix visuals
+    * Add sounds
+    * Add air bubbles particles
+    * Add air to player
 * Add crafting recipes